# docs/swagger.yaml
basePath: /
definitions:
...
  ent.CouponUsage:
    properties:
      coupon_code:
        description: 优惠券代码
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      discount_amount:
        description: 优惠金额(分)
        type: integer
      expire_at:
        description: 过期时间
        type: string
      id:
        description: ID of the ent.
        type: integer
      order_id:
        description: 订单ID
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: '状态: 0-未使用 1-已使用 2-已过期'
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      used_at:
        description: 使用时间
        type: string
      user_id:
        description: 用户ID
        type: integer
    type: object
...
  model.CouponUsageBatchDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
  model.CouponUsageBatchUpdateReq:
    properties:
      ids:
        items:
          type: integer
        type: array
      status:
        type: integer
    required:
    - ids
    type: object
  model.CouponUsageCreateReq:
    properties:
      coupon_code:
        type: string
      discount_amount:
        type: integer
      expire_at:
        type: string
      order_id:
        type: integer
      remark:
        type: string
      user_id:
        type: integer
    required:
    - coupon_code
    - discount_amount
    - expire_at
    - user_id
    type: object
  model.CouponUsageUpdateReq:
    properties:
      discount_amount:
        type: integer
      expire_at:
        type: string
      order_id:
        type: integer
      remark:
        type: string
      status:
        type: integer
      used_at:
        type: string
    type: object
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.PageResult-ent_CouponUsage:
    properties:
      records:
        items:
          $ref: '#/definitions/ent.CouponUsage'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/couponusage/batch/delete:
    delete:
      consumes:
      - application/json
      description: 批量删除优惠券使用记录
      parameters:
      - description: 优惠券使用记录批量删除请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.CouponUsageBatchDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 批量删除优惠券使用记录
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/batch/update:
    put:
      consumes:
      - application/json
      description: 批量更新优惠券使用记录
      parameters:
      - description: 优惠券使用记录批量更新请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.CouponUsageBatchUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 批量更新优惠券使用记录
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/create:
    post:
      consumes:
      - application/json
      description: 创建优惠券使用记录
      parameters:
      - description: 优惠券使用记录创建请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.CouponUsageCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.CouponUsage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建优惠券使用记录
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除优惠券使用记录
      parameters:
      - description: 优惠券使用记录 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除优惠券使用记录
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/list:
    get:
      consumes:
      - application/json
      description: 获取所有优惠券使用记录列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ent.CouponUsage'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取所有优惠券使用记录
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/page:
    get:
      consumes:
      - application/json
      description: 获取优惠券使用记录列表分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-ent_CouponUsage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取优惠券使用记录列表分页
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询优惠券使用记录
      parameters:
      - description: 优惠券使用记录 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.CouponUsage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询优惠券使用记录
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/search:
    get:
      consumes:
      - application/json
      description: 搜索优惠券使用记录
      parameters:
      - in: query
        name: coupon_code
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-ent_CouponUsage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 搜索优惠券使用记录
      tags:
      - 优惠券使用记录
  /api/v1/couponusage/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新优惠券使用记录
      parameters:
      - description: 优惠券使用记录 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 优惠券使用记录更新请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.CouponUsageUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.CouponUsage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新优惠券使用记录
      tags:
      - 优惠券使用记录
...
swagger: "2.0"